First implementation of all the tiles
The ways have been used to implement tiles. The first is to simply have classes implementing the Tile interface such as 'ExitTile' and 'Wall'. The second is to have an 'ItemTile' which can contains one Item which changes its state. For example it can be just an emptyTile if there is no item or a playerTile if there is a player. This allow more modularity because the player (and other items such as treasures and keys) do not always have to be on a basic tile but in the future they could be on ice or trampolines