Change how player position is store #2
Now the coordinates of the tile where the player is located are stored as a Point in the player class and the maze stores a reference to the Player object instead of its coordinates. This allow the maze to access the player inventory easier (which will be needed by the application).