Change Logic, entity can now interact with Items through action
The concept of actions linked to actions is now been introduced. This allow a greater freedom in term of what an option can do. In this way it is also easier to record when special action occurs in the game. This is an evolution of the pickup method which was limited to Items that were 'Collectable'. With this change the code is also more robust since Keys are now been removed when the action of removing the block is executed instead of when it was checked if the player could remove the block. Hence removing possible side effects.